refactor(scheduler): wire ServiceChecker + RetentionManager, extract rules.go (#94)
